requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
scipy==1.13.0
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
reportlab==4.2.0
matplotlib==3.9.0
pyyaml==6.0.1
pytest==8.2.0
Pillow==10.3.0
